chore: restate-sdk 0.4 -> 0.10 + canonical Clorinde + schema regen (#116)
* chore(clorinde): regenerate bindings against current schema

* chore(restate-sdk): upgrade 0.4 -> 0.10

Restate-sdk 0.10 introduces a PayloadMetadata trait that must be implemented
for any non-primitive type used directly as a handler input/output in a
#[restate_sdk::workflow] or #[restate_sdk::object] trait. Extended the
impl_restate_serde_via_json! macro in restate/types.rs to also emit
impl PayloadMetadata so existing call sites pick it up automatically.

Cleaned up the stale "restate-sdk 1.10 is the compatible range" comment
above the bytes dep (the "1.10" referred to the wrong crate). bytes = "1"
is kept because restate-sdk 0.10 still uses bytes::Bytes in its serde
trait method signatures.

The other three SDK consumers (service.rs, http_endpoint.rs, mod.rs) are
unchanged - the Context API, prelude, and HTTP wiring carried over cleanly.

This resolves the sha2 stable/pre-release conflict that previously blocked
clorinde from regenerating postgres-protocol 0.6.11 bindings on main.

* chore: regenerate schema.pg.sql.generated

Regenerated against canonical Postgres state via the schema.pg.sql.generated-fresh
artifact from CI run 25776599469 on PR #114. The previous version on main was
missing changes to correlation_topics and other tables that canonical Postgres
already has, so the schema-fresh-verify job was failing on any queries/**
touch.

* chore(clorinde): regenerate bindings against current schema

* chore: refresh schema.pg.sql.generated against latest canonical

The previous regen used the artifact from PR #114's run (25776599469),
which was already stale: a newer qontinui-web alembic migration added
identity_hash text to coord.tasks plus a partial unique index on
(plan_id, identity_hash). Picked up from this PR's own
schema-fresh-verify run (25794719690) artifact.

---------

+++ src-tauri/clorinde/src/queries/graph_ops.rs
@@ -35,7 +35,7 @@ pub struct InsertStepFindingLinkParams<
     pub step_index: i32,
     pub finding_id: T4,
     pub link_type: T5,
-    pub confidence: f32,
+    pub confidence: f64,
 }
 #[derive(Debug)]
 pub struct GetFindingsForStepParams<T1: crate::StringSql, T2: crate::StringSql> {
@@ -200,7 +200,7 @@ pub struct StepFindingLinkRow {
     pub step_index: i32,
     pub finding_id: String,
     pub link_type: String,
-    pub confidence: f32,
+    pub confidence: f64,
     pub created_at: chrono::DateTime<chrono::FixedOffset>,
 }
 pub struct StepFindingLinkRowBorrowed<'a> {
@@ -210,7 +210,7 @@ pub struct StepFindingLinkRowBorrowed<'a> {
     pub step_index: i32,
     pub finding_id: &'a str,
     pub link_type: &'a str,
-    pub confidence: f32,
+    pub confidence: f64,
     pub created_at: chrono::DateTime<chrono::FixedOffset>,
 }
 impl<'a> From<StepFindingLinkRowBorrowed<'a>> for StepFindingLinkRow {
@@ -1246,17 +1246,17 @@ impl InsertWorkflowVersionStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-        T6: crate::StringSql,
-        T7: crate::StringSql,
-        T8: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+    T6: crate::StringSql,
+    T7: crate::StringSql,
+    T8: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
@@ -1413,7 +1413,7 @@ impl InsertStepFindingLinkStmt {
         step_index: &'a i32,
         finding_id: &'a T4,
         link_type: &'a T5,
-        confidence: &'a f32,
+        confidence: &'a f64,
     ) -> Result<u64, tokio_postgres::Error> {
         client
             .execute(
@@ -1432,14 +1432,14 @@ impl InsertStepFindingLinkStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
@@ -1640,18 +1640,18 @@ impl InsertStepProvenanceStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-        T6: crate::StringSql,
-        T7: crate::StringSql,
-        T8: crate::StringSql,
-        T9: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+    T6: crate::StringSql,
+    T7: crate::StringSql,
+    T8: crate::StringSql,
+    T9: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
@@ -1861,15 +1861,15 @@ impl InsertPipelineEventStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-        T6: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+    T6: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
@@ -2046,16 +2046,16 @@ impl InsertRuleInfluenceStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-        T6: crate::StringSql,
-        T7: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+    T6: crate::StringSql,
+    T7: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
@@ -2223,16 +2223,16 @@ impl UpsertCrossRunPatternStmt {
     }
 }
 impl<
-        'a,
-        C: GenericClient + Send + Sync,
-        T1: crate::StringSql,
-        T2: crate::StringSql,
-        T3: crate::StringSql,
-        T4: crate::StringSql,
-        T5: crate::StringSql,
-        T6: crate::StringSql,
-        T7: crate::StringSql,
-    >
+    'a,
+    C: GenericClient + Send + Sync,
+    T1: crate::StringSql,
+    T2: crate::StringSql,
+    T3: crate::StringSql,
+    T4: crate::StringSql,
+    T5: crate::StringSql,
+    T6: crate::StringSql,
+    T7: crate::StringSql,
+>
     crate::client::async_::Params<
         'a,
         'a,
